# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Timekeeping and Helpdesk API
  version: 1.0.0
paths:
...
  /api/v1/auth/login:
    post:
      operationId: auth_login_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Login'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      tags:
      - api
  /api/v1/auth/logout:
    post:
      operationId: auth_logout_create
      responses:
        '204':
          description: No response body
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/auth/refresh:
    post:
      operationId: auth_refresh_create
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      tags:
      - api
  /api/v1/change-password:
    post:
      operationId: identity_change_password_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordChange'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordChange'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
...
  /api/v1/me/:
    get:
      operationId: identity_me_retrieve
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfUser'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
    patch:
      operationId: identity_me_partial_update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Profile'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfUser'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
...
  /api/v1/users/:
    get:
      operationId: users_list
      parameters:
      - in: query
        name: is_active
        schema:
          type: boolean
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: q
        schema:
          type: string
      - in: query
        name: role
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
    post:
      operationId: users_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedUserResult'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/users/{user_id}/:
    get:
      operationId: users_retrieve
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
    patch:
      operationId: users_partial_update
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Profile'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/users/{user_id}/reset-password:
    post:
      operationId: users_reset_password_create
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordResult'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/users/{user_id}/role:
    patch:
      operationId: users_role_partial_update
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Role'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
  /api/v1/users/{user_id}/status:
    patch:
      operationId: users_status_partial_update
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Status'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Status'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Status'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityError'
          description: ''
      security:
      - bearerAuth: []
      tags:
      - api
components:
  schemas:
    AccessResponse:
      properties:
        access:
          readOnly: true
          type: string
      required:
      - access
      type: object
...
    IdentityError:
      properties:
        details:
          additionalProperties: {}
          readOnly: true
          type: object
        error:
          readOnly: true
          type: string
        error_code:
          readOnly: true
          type: string
        message:
          readOnly: true
          type: string
        request_id:
          format: uuid
          readOnly: true
          type: string
      required:
      - details
      - error
      - error_code
      - message
      - request_id
      type: object
...
    Login:
      properties:
        password:
          type: string
          writeOnly: true
        username:
          type: string
      required:
      - password
      - username
      type: object
    LoginResponse:
      properties:
        access:
          readOnly: true
          type: string
        capabilities:
          items:
            type: string
          readOnly: true
          type: array
        is_active:
          readOnly: true
          type: boolean
        must_change_password:
          readOnly: true
          type: boolean
        role:
          readOnly: true
          type: string
      required:
      - access
      - capabilities
      - is_active
      - must_change_password
      - role
      type: object
...
    PasswordChange:
      properties:
        current_password:
          type: string
          writeOnly: true
        new_password:
          type: string
          writeOnly: true
      required:
      - current_password
      - new_password
      type: object
...
    Profile:
      properties:
        email:
          format: email
          nullable: true
          type: string
        full_name:
          type: string
        phone:
          nullable: true
          type: string
      type: object
...
    UserPage:
      properties:
        count:
          readOnly: true
          type: integer
        next:
          nullable: true
          readOnly: true
          type: string
        previous:
          nullable: true
          readOnly: true
          type: string
        results:
          items:
            $ref: '#/components/schemas/AdminUser'
          readOnly: true
          type: array
      required:
      - count
      - next
      - previous
      - results
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http